Viewport proxy support
Change the Import UI to display a boolean "Viewport Proxy".

If set (default), the low res proxy will be visible in the viewport. If
not, the high res model will be visible in the viewport with only a
wireframe of the proxy visible. The proxy object is not visible for
rendering, the high res model is visible for rendering.

The proxy object is setup as the parent object with the high res object
as a child. When selecting the proxy and manipulating it (move, scale)
the child object follows the parent. To ensure consistency, the high res
model is not selectable - all manipulations must be made to the proxy.

Signed-off-by: Darren Hart <dvhart@infradead.org>