Distinguish proxy and full materials
The Laubwerk SDK provides the same plant material name for the trunk and
branch materials for the proxy and the full models. Since we only
generate diffuse solid color materials for the proxy object, change the
blender name of the material by appending _proxy when putting and
getting the blender material.

Signed-off-by: Darren Hart <dvhart@infradead.org>