Update plant related naming scheme
Use the plant.name to refer to all meshes and objects. Do not attempt to
get more specific with variant, age, season, as those terms are not
sufficient to describe the model anyway (would need all the meta data to
do that). Instead, use the plant name and let Blender append the typical
.001, .002 numerical disambiguating suffixes. This is consistent
with the Laubwerk plugin for SketchUp.

Refer to the viewport and render objects using those terms in the
variable names and the mesh and object name properties.

For proxy mesh materials, use the same plant name with foliage or wood
as a suffix.

As it appears to be typical to use spaces in naming of objects within
Blender, do not attempt normalize the plant.name by replacing spaces or
changing case. Use what is reported by the Laubwerk SDK.

Signed-off-by: Darren Hart <dvhart@infradead.org>