Always create a viewport and render object
Always generate a separate viewport and render object. If viewport_proxy
is set, create a very low poly convex hull, otherwise create a low
detail model with 30% of the render leaf density.

Reduce the complexity of the object and material functions by dealing
with the UI variables entirely within the load function.

Contain the proxy special case handling to the object creation function
and minimize the if blocks.

Remove the leftover code for wireframe and boundary rendering.

Signed-off-by: Darren Hart <dvhart@infradead.org>