Use the LaubwerkDB for plant data rather than the SDK
The LaubwerkDB provides caching for the known Plant data, speeding up
the Import process by replacing SDK queries with python dict lookups.
Simplify the logic significantly by eliminating the watch thread.

Add Addon Preferences to specify the Laubwerk install path and Update
Database from there.

Update the import panel to instruct the user to update the database via
the Addon Preferences when a plant filename is not found in the
database. Simplify the draw logic.

Signed-off-by: Darren Hart <dvhart@infradead.org>