Remove unused imports, classes, and properties
We were not conditionally passing the level and thickness level of
detail settings. Rather than add the logic to do so, use reasonable
defaults.

Simplify the panel draw() by removing the subrow active logic based on
these bool properties.

Remove the unsed lbwPanel class as well as the unused reset to defaults
block.

Signed-off-by: Darren Hart <dvhart@infradead.org>