Use a fixed db path, global plants and sdk paths, local enum items
Use a Blender user path to store the db in a fixed location (user addons
dir matching the addon name - which is probably where the addon is
installed - but may not be if it is stored in the system addons).

Use global vars for plants and sdk paths, making it possible to manage
the db lifecycle automatically - initialize it at register time, update
it in preferences.

Fix a model/season bug resulting in rendering errors. The lists were not
being reset before the new plant models were added. The global enum
items are no longer necessary with the LaubwerkDB, eliminate them in
favor of local db lookups.

Signed-off-by: Darren Hart <dvhart@infradead.org>