Group plant models into collections
The parent/child object grouping of the viewport and render object is
problematic for a couple of reasons:

- Deleting the parent from the viewport leaves the child dangling and
  not visible or selectable in the viewport.
- Selecting the parent as an object to emit in a particle system only
  scatters the parent so the render is empty.

Address this by:

- Create a collection for each plant and place the objects in it
- Place the plant collection in a top level collection ("Laubwerk")
- Create a collection instance from the top level collection to place in
  the active collection.

The collection instance is a "single object" within the viewport and can
be deleted with the standard select and 'x'.

When emitting objects in a particle system, choose "collection" and
select the name of the plant collection. This will emit instances which
contain both the viewport and render visible meshes.

Signed-off-by: Darren Hart <dvhart@infradead.org>