Confirm scale and performance questions
Blender appears to use units of meters for meshes created with
from_pydata, regardless of scene units. There is no need to configure
this scale factor as it will always be 0.01. Make the code reflect this
and simplify the vertices list generation.

The universally accepted way to use smooth shading on an object is to
loop through the faces and set them all to smooth. Unfortunately, there
does not a appear to be a way to do this as the list is being imported,
forcing us to repeat the O(N) operations.

Tests show that the scale multiply and the face smoothing have less
impact than the standard variance in import times for the Acer Palmatum
test plant (it's all within the noise).

Signed-off-by: Darren Hart <dvhart@infradead.org>
(cherry picked from commit 4c7807a9a8dd50dc22ae6bfdf7ae71d6faf891a1)