Add Python 3 SDK dependency notice
Laubwerk has declined to share the development version of the Python 3
SDK. Without it, users cannot experiment with this Add-on. Update the
README to make this dependency explicit and remove the instruction to
request early access to the SDK.

Note that this is temporary. Once Laubwerk releases the Python 3 SDK, it
will be packaged with all their plant packs and all Laubwerk users will
be able to use this Add-on.

Signed-off-by: Darren Hart <dvhart@infradead.org>