Specify python argument to LaubwerkDB from lbwdb_write
Commit b3da6243bc80 "Refactor locale handling" added the locale argument
to the LaubwerkDB constructor, changing the position of the python
argument. This resulted in calls to rebuild the database to be called
using the default sys.executable, which is the Blender app when run
inside Blender. Attempting to run the python script from the Blender
binary resulted in unexpected output being passed into json.loads(), and
json errors being reported in the log.

Update the call from lbwdb_write to explicitly specify python=.

Fixes: b3da6243bc80 Refactor locale handling

Signed-off-by: Darren Hart <dvhart@infradead.org>