Major README update
The Laubwerk 1.0.32 Plant Packs now include the Python3 SDK. Remove the
warning about availability and update the requirements to specify 1.0.32
or later.

Add a clear disclaimer stating Thicket is a community project not
affiliated with Lauberk GmbH - but also giving them credit for making
this possible and cooperating with the project.

Cleanup and refactor the introduction and installation sections.

Add a usage and collections section, as well as a blurb about how to
report issues.

Update several images.

Signed-off-by: Darren Hart <dvhart@infradead.org>