lbwdb: Avoid plant parsing hang on Windows
Building the database would hang consistently on Windows on Acer_capestre
from the Plants Kit Freebie. sub.poll() would return None even after
the process had completed. I was not able to determine why - but there
are many reports of similarly frustrating behavior out of subprocess on
Windows.

By avoiding poll() and wait(), and using communicate() directly, the
problem goes away. While this is not as efficient and may leave several
processes completed and ready while waiting for the next in the list to
complete, the total time to complete parsing the 56 plants in my library
is unchanged.

Signed-off-by: Darren Hart <dvhart@infradead.org>