Fix Rebuild DB failure when not run from command line
Rebuild DB would fail when the Laubwerk SDK path was not in the
PYTHONPATH environment (for this was if I launched it normally instead
of from the command line).

Pass sdk_path along to lbwdb_plant.py and add it to the sys.path if it
is not already there. This approach proved more portable between Mac and
Windows than setting the env path explicitly via the Popen call (which
worked on Mac and failed on Windows - which already appears to the the
Laubwerk SDK int he PYTHONPATH).

Signed-off-by: Darren Hart <dvhart@infradead.org>

sdk_path only

Signed-off-by: Darren Hart <dvhart@infradead.org>