Use package relative imports
In the thicket package, use . instead of thicket to refer to the
package. This allows for the thicket name to change (such as with a
version name) without needing to change the code.

This allows for creating versioned Blender Add-on packages, such as
thicket_020.zip. Thanks to Anthony Kavasis for the recommendation.

Signed-off-by: Darren Hart <dvhart@infradead.org>