Force reload after importing individual plant to db from filebrowser
Because the plant was not in the database, the model_id and
model_season properties are empty. We need to force reloading them or
executing import the import after it is added to the DB will result in
failure - not finding the model_id in the empty enum.  Force this by
setting oldpath to the empty string, which will trigger draw() to treat
this as a new file.

Signed-off-by: Darren Hart <dvhart@infradead.org>