lbwdb: Throttle rebuild database operator
The current build() routine spawned one process per plant, regardless of
system capabilities. For large numbers of plants, this can result in
consuming a lot of memory while unpacking the plants, resulting in
swapping, which severely degrades performance (10x in one of my tests).

Implement a FIFO queue for jobs and limit the number of jobs to the
number of available CPUs (this is the number of CPUs available to the
process as reported by os.sched_affinity) or 4 if that returns None.

This has a 20% slowdown in processing on a 6 core hyperthreaded / 32G
system (21 seconds longer), but a 87% speedup on a more modest 4 core /
16G system (8.7 minutes faster). A worthy tradeoff.

Signed-off-by: Darren Hart <dvhart@infradead.org>