Add Thicket Property Editor
Identify Thicket objects by adding custom property groups to the
Collection type, and populate it on Thicket plant template collections.
This will make it possible to distinguish them from other objects in the
scene and manipulate them. Use a "unique" GUID in a "magic" field out of
an abundance of caution.

collection.thicket: model properties
collection.thicket_shadow: copy of thicket to use in the UI to buffer
                           changes until the user confirms them.

The Thicket Properties panel allows the user to change the properties
(backed by thicket_shadow) and either Reset them to the originals (copy
thicket back to thicket_shadow) or Update the plant, which will update
the template plant (the one in the Thicket collection) which will in
turn update all the plants instanced off the template.

Reduce the visual clutter of the Thicket Properties panel and make it a
little more blender-like by adding a small button to the right of the
plant name with number of instances of the same template plant, clicking
on it makes it unique (much like the 'make single instance' operators in
the outliner.

Add thicket_utils file and introduce a Delete Plant operator which will
delete all resources used by a Thicket Plant if it is the last instance
of the plant template.

Signed-off-by: Darren Hart <dvhart@infradead.org>