Workaround model.get_mesh qualifier keyword
The SDK documents the keyword as qualifier_name, the library expects
qualifierName (leftover from the Python2 code), and will likely
standardize on qualifier. Pass both qualifierName and qualifier so it
works today, will work in the future without changes, and we can drop
qualifierName once the API is finalized.

Use the term qualifer throughout for consistency.

Signed-off-by: Darren Hart <dvhart@infradead.org>