Update Plant Properties and Import Panels
Update the UI for the Plant Properties in both the 3D View and the
Import to be explicit, clear, and as minimal as possible.

Use a common draw routine for both panels, relying on a common set of
properties for both contexts. Unfortunately I have not found a way to
create these properties in only one place, as one depends on the active
object, and the other is preparing to create an object.

Signed-off-by: Darren Hart <dvhart@infradead.org>