Add camelCase arguments to get_mesh
The current SDK is using camelCase arguments, but is documented as using
underscore case. Use both for compatibility while the SDK is updated.

We were previously passing qualifier and qualifierName, so the season
was reflected in the generated mesh, but all the other level of detail
arguments were being ignored. This gets them working again.

Signed-off-by: Darren Hart <dvhart@infradead.org>