Fix Laubwerk Install Path failure on new install
A new install of the Addon resulted in in failure due to the logic in
thicket_init allowing it to continue if the path was not set.

A new bug cropped up in recent Blender updates since I had tested this
last in that Blender relative paths are stored in the StringProperty for
the DIR_PATH subtype. Address this by converting it to absolute in the
update callback. Set the preperty directly in the dictionary to avoid an
infinite update loop.

Signed-off-by: Darren Hart <dvhart@infradead.org>