Introduce Bi-modal Thicket N Panel with Gallery View
Extend the N Panel to allow for changing the plant model for the active
plant (and it's siblings) as well as add a new plant.

Clicking Add Plant or Change plant switches to a gallery view to allow
the user to select one of the installed Laubwerk plants, then switches
to the edit view to configure the plant, before updating or adding.

Manage the states through module globals. Attempting to use properties
attached to the context.window_manager resulted in too many limitations
with when they can be set (not in draw()), which overly complicated
management of the state machine.

The following states are implemented for the Thicket N Panel:
  - view
  - edit
  - add
  - select
  - select_add

Signed-off-by: Darren Hart <dvhart@infradead.org>