Update README with N Panel UI
Update the README Usage section with instructions for the N Panel UI.
Move the Import Menu to Deprecated.

Drop Laubwerk Python3 SDK in favor of Laubwerk Python SDK as the Python
version is not relevant to the end user (only the Laubwerk version
1.0.32 is).

Remove the File->Import instructions (it's been removed from the
Add-on) in favor of the N Panel UI.

Update the screenshots using the latest UI and a consistent size and
capture method.

Signed-off-by: Darren Hart <dvhart@infradead.org>