Refine N Panel Modes and State Machine
Move state machine logic to a function and use it consistently with
operators which change the UI mode.

Only show Add and Delete in View mode.

Move Edit below the image to group it with the properties you are
enabling.

Move Make Unique below Change Plant in Edit Mode only since it only
makes sense to make a plant unique if you are going to edit it.

Switch to VIEW mode after confirming property choices with Add or
Update, since Cancel does not make as much sense after confirming those
changes.

Signed-off-by: Darren Hart <dvhart@infradead.org>