Add instruction to verify install path contains Plants and Python directories
If the Install Path fails the test for a Plants and Python directories,
give the user something to check with a message, also in red, that the
install path should contain these two directories.

Only change the enabled state of the Rebuild Database button. No need to
gray out the SDK version and database plant count since those message
change to reflect the add-on's state already.

Signed-off-by: Darren Hart <dvhart@infradead.org>