Key plant lookup on plant name rather than filename
Make the Thicket object properties robust to changes to the filesystem
layout of the Laubwerk Plant Kits. Use name instead of filepath to
identify the plants. If the installation changes, rebuilding the DB is
sufficient to find all the plants again - whereas storing the filepath
in the object loses the identity of the plant if the file is removed or
moved.

Use pop() instead of property_unset() as it is resilient to the filepath
being removed from the property group if it exists in existing scenes.

Signed-off-by: Darren Hart <dvhart@infradead.org>