Add ThicketStatus enum and handle database schema upgrades
Breaking database schema changes currently go undetected, resulting in
failure.

Add a ThicketStatus enum to manage a set of startup states to provide
useful information to the user if something fails along the way.

If an old database schema is detected, remove the database and create an
empty one, as though we were loading the add-on for the first time, and
set the status to REBUILD_DB. This is destructive, but the db is not
intended to be user modifiable and is always recreatable from the
Laubwerk plant library.

Add logic to the preferences panel to rebuild the database.

Add logic to the N Panel to rebuild the database if needed.

This makes it clear what is needed to recover and doesn't enter a
possibly time consuming rebuild process during init which the user is
likely to cancel.

Signed-off-by: Darren Hart <dvhart@infradead.org>