thicket_db: key database on plant name and update to schema version 2
Now that we are no longer using the file->import mechanism, the user
interacts with Thicket via the plant names, not their filenames. Most
get_plant() calls have the name, not the filename. Update the database
to key on the plant name now while we're in alpha status before the
database schema changes are a bigger deal to users (the add-on already
handles these changes gracefully).

Signed-off-by: Darren Hart <dvhart@infradead.org>