Handle unknown plants
If a plant has the THICKET_GUID but is not found in the database,
don't crash, but notify the user in the N panel and provide the operator
to rebuild the database.

Avoid similar problems in the general case when the plant count is 0. To
be useful, the user will have to rebuild the database, so make that
obvious and easy to find.

Signed-off-by: Darren Hart <dvhart@infradead.org>