Do not rely on "name in sys.modules"
If two versions of thicket are installed, the required modules may
appear in sys.modules, but the name may not be defined in the thicket
global namespace, resulting in a name not found exception.

sys.modules lists the loaded modules, which may or may not have been
imported into the current namespace. Drop the name in sys.modules test
and try to import all modules each time thicket_init is called.

Signed-off-by: Darren Hart <dvhart@infradead.org>