Add "Full Geometry" to viewport modes
Add the option to use the same geometry for the viewport as in renders.
When selected, only load one mesh from the Laubwerk SDK (rather than
having the viewport object be the parent of the render object).

This relies on lbw_plant.name instead of the obj_viewport.name
throughout, but this doesn't produce meaningfully different results in
the final object names.

Fixes: Issue #16: Viewport: Use render model

Signed-off-by: Darren Hart <dvhart@infradead.org>