Add Render level of detail control for Proxy or Full Geometry
In some cases, users may wish to use the proxy model even in full
renders. Provide a drop down box Level of Detail to select PROXY or FULL
for the Render.

If the Render level of detail (LOD) is set to PROXY, force the the
viewport LOD to PROXY and disable the control in the UI.

This enforces consistency with the logic in thicket_lbw and avoids
creating duplicate proxy objects.

The only time a viewport object is created in addition to the render
object is when the viewport is PROXY or PARTIAL AND the render is FULL,
otherwise the viewport and the render both use the render object.

Fixes: Issue #15: Add proxy to render detail

Signed-off-by: Darren Hart <dvhart@infradead.org>