Update Plant Properties UI to look more like other Laubwerk Players
Leaf Density is not a Level of Detail setting, but rather a property of
the plant like model and season. Move Leaf Density out of LOD.

Sort LOD properties by priority (adjust this first, then this.... and
this as a last resort).

Move plant name above the thumbnail. By moving the name and botanical
name out of draw_props and moving above the thumbnail, the labels are
always "enabled" (not grayed out) regardless of VIEW or EDIT mode,
making them easier to read.

Signed-off-by: Darren Hart <dvhart@infradead.org>