Make branching level and thickness optional parameters
Allow the the user to leave these parameters to the Laubwerk SDK.

Simplify calling import_lbw and get_mesh by consolidating calls to
import_lbw to a wrapper in ThicketPropGroup, and eliminate the
as_keywords() method.

Pass a mesh_args parameter only for those arguments used in the get_mesh
call.

Keep the ThicketPropGroup property names the same and map them to the
Laubwerk SDK keyword names rather than changing them for now - this
avoids breaking existing usage where the existing names are used. We may
align these names prior to the 1.0 release.

Fixes: Issue #18: Make Max Branching Level and Min Branch Thickness optional

Signed-off-by: Darren Hart <dvhart@infradead.org>