Specify Laubwerk Plants and Python paths separately
The Laubwerk Plant Kits allow the user to specify different locations
for the Plants and the Python Extension. Allow for this by specifying
the two paths individually in the Addon Preferences, rather than
assuming a single top level Laubwerk install path.

To be able to check the two paths and provide meaningful error messages
in the preferences panel, convert the ThicketStatus object from an enum
to a simple object with several bool members. Change the lbw_sdk_valid
semantics to mean the laubwerk module was successfully loaded. Drop the
rebuild_db state as this is equivalent to the ready state.

Simplify installation for the common case by providing default paths for
both Windows and Mac (Darwin) users via the platform.system() call.

Fixes: Issue #20: Separate Laubwerk Plants and Python folders

Signed-off-by: Darren Hart <dvhart@infradead.org>