PEP8 Linter: Use "python3 -m" and update arguments
Prefix flake8 with "python3 -m" to work if pip is used to install flake8
rather than the distro package manager.

Update the ignore list and drop the select list to catch more errors.
Specifically ignore F722,F821 as all the Blender *Property declaration
are flagged (seems to be a false positive).

Add terminal description of what is being run to make the output easier
to interpret.

Drop the --count argument as it doesn't offer anything useful.

Update the GitHub action to match.

Signed-off-by: Darren Hart <dvhart@infradead.org>