Support two sided leaf materials
Some Laubwerk materials have two diffuse (color) textures, one for each
side of a leaf, for example. Check if the material has two distinct
sides and if so load the second image and use an RGBMix shader
controlled by a Geometry shader backface output to render each side of
the leaf with a different image.

Fixes: Issue #23: Support two sided leaf materials

Signed-off-by: Darren Hart <dvhart@infradead.org>