Simplify geometry creation and fix normals
The existing code scaled and rotated the point data from Laubwerk to map
it to Blender's scale and orientation, adding additional logic,
intermediate variables, and lots of multiplication within the Python
code. It also resulted in the normals being flipped.

Reduce the logic required by using the points and polygons directly from
Laubwerk and perform the scale and rotation using a single matrix
operation. This also results in the normals pointing in the right
direction.

This reduces the amount of code and reduces the total plant import time
slightly (~4% in my tests).

Signed-off-by: Darren Hart <dvhart@infradead.org>