Initial batch edit support
Add batch properties to the ThicketPropGroup for:
  - batch_mode
  - batch_name
  - batch_model
  - batch_qualifier
  - batch_use_lod

These inform the logic if operating in batch mode when the context is
not available and whether or not to change certain properties.

If batch_name is empty, use the original plant. If batch_use_lod is
False, use LOD settings from the original plant.

The model and qualifier properties are new Enums which include the
default Laubwerk models and seasons and an additional UNCHANGED value
which allows for a "--" to be displayed in the Edit N Panel entries.

When updating a list of plants, only update any given template once,
regardless of how many instances are selected. Restore the selected
objects and the active objects before finishing.

Update labels such as "Edit Plant" to "Edit" to avoid having to deal
with singular and plural terms.

Use a preview with multiple trees for batch editing.

Fixes: Issue #14: Bulk plant edits

Signed-off-by: Darren Hart <dvhart@infradead.org>