Support make unique in batch mode
Update the make_unique operator to loop through the selected Thicket
instance objects, rather than just the active object.

Move the button up with Add and Delete so the user doesn't have to Edit
a plant in order to make it unique.

Update the label to display the total number of siblings (but only
counting each instance once).

Signed-off-by: Darren Hart <dvhart@infradead.org>