Support Laubwerk 1.0.34 materials
The Laubwerk Python Bindings v1.0.34 made some significant changes to
the material model. This patch makes the minimal changes necessary to
accomodate these changes, mainly property name changes and an update to
how subsurface scattering and transmission is determined.

Keep the old lbw_to_bl_mat implementation as lbw_to_bl_mat_1033 and use
it when 1.0.33 or earlier is detected (nothing earlier is supported).
Log a warning that 1.0.33 support is deprecated and will be removed in
future releases.

In both implementations, remove the subsurface color as it was not being
used since the subsurface multiplier is 0 for thin-shell materials (and
solid-volumes have not been supported).

Fixes: Issue #26: thicket throws python errror when trying to add plant

Signed-off-by: Darren Hart <dvhart@infradead.org>