Use translucency rather than transmission
The subsurface properties for this shell materials should be modeled
with a Translucent BSDF rather than the Principled BSDF Transmission
property. This is not a complete solution, but more closely approximates
the intent of these properties than the previous material node network.

Related to Issue #35: Translucency seems to be missing

More work is needed as this is a particularly complex problem
apparently:
https://devtalk.blender.org/t/diffuse-translucency-way-too-difficult-to-set-up-in-cycles/807

Signed-off-by: Darren Hart <dvhart@infradead.org>