logging: Use thicket specific Logger and a FileHandler
Thicket was using the root Logger which contradicts the Blender guidance
on logging [1].

Create a Thicket specific Logger and attach a FileHandler so the log
always goes to a file, making it easier for users to find it without
having to run Blender from the command line and simplifying error
reports.

Propagate the messages so advanced users can configure Blender logging
according to their preferences.

Update thicket_db to avoid logging when called as a subprocess by moving
the Parsing message into the build function.

The new log file (thicket.log) is found in the same folder as the add-on
itself.

1. https://code.blender.org/2016/05/logging-from-python-code-in-blender/

Fixes: Issue #38: Use a module specific logger

Signed-off-by: Darren Hart <dvhart@infradead.org>