Retain plant properties when canceling plant selection
If changes are made to plant properties in the Add or Edit modes and the
use then switches to the Select_Add or Select states, and then cancels
that operation, the UI returns to View state instead of the previous
state, and the property edits are lost. A user can reasonably expect the
Cancel button to take them back to the previous view of the UI with
their edits preserved.

Refactor select_plant() out of the select_plant operator so it can be
called from the change_mode operator.

If there is no UI ThicketPropGroup configured when the user presses Add
Plant, select the first plant from the DB.

Update the UI state transitions such that Add Plant goes directly to ADD
(instead of SELECT_ADD) and Cancel while in SELECT or SELECT_ADD return
to EDIT or ADD respectively.

Fixes: Issue #44: Retain plant properties when canceling plant selection

Signed-off-by: Darren Hart <dvhart@infradead.org>