Do not draw plant props if no plants are selected
When in VIEW mode, if no plants are selected, but the active object is a
plant, the UI will display its properties - but attempting to complete
an edit operation will have no effect as it is now based on selection in
order to accommodate batch processes.

Update the N-Panel draw method to exit after drawing the Add Plant
operator if there are no plants selected. This mirrors the previous
intended behavior and provides a more intuitive user experience.

Signed-off-by: Darren Hart <dvhart@infradead.org>