Reuse viewport and render meshes when unchanged
When editing a plant, both meshes will be regenerated whether or not the
changes require it. For example, if the viewport is changed from FULL to
PROXY, the render object will also be reloaded, causing the operation to
take much longer than necessary (several seconds longer).

Update the thicket_lbw.import_lbw function to accept viewport and render
objects to reuse. If they are passed it, use them instead of
regenerating the meshes. Be sure to update all visibility and parenting
properties and do not assume defaults (e.g. in the case of changing from
PROXY:FULL to FULL:FULL the render object will be reused, but it should
it needs to be made explicitly visible as it was previously not shown in
the viewport.

Update ThicketPropGroup to have an eq_lod() function to compare only
level of detail properties. Only attempt to reuse objects if:
* Plant, model, qualifier are unchanged
* LOD settings are unchanged
  * render: type is unchanged
* viewport: remains a proxy
* render: becomes a proxy and viewport was/is a proxy

Note: There is no way to change the plant properties such that a LOW
viewport mesh can be reused. If the LOD changes, the LOW mesh needs to
change. If the Render mode changes to PROXY, the viewport is forced to
PROXY.

Fixes: Issue #42: Optimize plant update by not reloading unnecessarily

Signed-off-by: Darren Hart <dvhart@infradead.org>