Move ThicketPropGroup setup out of thicket_lbw.import_lbw
The current code converts the ThicketPropGroup into mesh_args and sends
them to import_lbw, which then recreates the ThicketPropGroup from
mesh_args.

Avoid this step and simplify the logic by copying the ThicketPropGroup
from the operation in __init__.py rather than rebuilding it one property
at a time in thicket_lbw.import_lbw. This is more in keeping with
thicket_lbw.py's scope being to conver lbw objects to blender objects,
and not with managing the Thicket specific meta data.

Signed-off-by: Darren Hart <dvhart@infradead.org>