thicket_lbw: Update variant access for Laubwerk API v1.0.51
v1.0.51 of the Laubwerk API removes the name field from the ModelVariant
object, forcing us to index the variants[] array by discovering the
index by walking the params.variants.enum.options dictionary for a
match.

This is a minimal hack to get it working, but this open coded params
walk should be encapsulated in a function outside the logic blocks.

Signed-off-by: Darren Hart <dvhart@infradead.org>